+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/graphql/AbstractEthGraphQLHttpServiceTest.java
@@ -29,7 +29,6 @@ import org.hyperledger.besu.ethereum.core.DefaultSyncStatus;
 import org.hyperledger.besu.ethereum.core.Synchronizer;
 import org.hyperledger.besu.ethereum.core.Transaction;
 import org.hyperledger.besu.ethereum.eth.EthProtocol;
-import org.hyperledger.besu.ethereum.eth.manager.EthScheduler;
 import org.hyperledger.besu.ethereum.eth.transactions.PendingTransaction;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPool;
 import org.hyperledger.besu.ethereum.mainnet.ValidationResult;
@@ -146,8 +145,7 @@ public abstract class AbstractEthGraphQLHttpServiceTest {
                 GraphQLContextType.SYNCHRONIZER,
                 synchronizerMock,
                 GraphQLContextType.GAS_CAP,
-                0L),
-            Mockito.mock(EthScheduler.class));
+                0L));
     service.start().join();
 
     client = new OkHttpClient();